tools/mtp-conformance/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mtp-conformance"
version = "0.5.0"
description = "Conformance runner for MTP (Methodology Transfer Protocol) tooling and fixtures"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Lubor Fedák", email = "lubor@codemoravia.com" },
]
keywords = ["mtp", "conformance", "fixtures", "validation", "runtime"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.0",
    "jsonschema>=4.20.0",
    "pyyaml>=6.0",
    "mtp-lint",
    "mtp-run",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]

[project.scripts]
mtp-conformance = "mtp_conformance.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mtp_conformance"]
